LayoutTests/imported/w3c/web-platform-tests/content-security-policy/reporting/multiple-report-policies-expected.txt

CONSOLE MESSAGE: [Report Only] Refused to load ftp://blah.test/ because it does not appear in the img-src directive of the Content Security Policy.
CONSOLE MESSAGE: [Report Only] Refused to load ftp://blah.test/ because it does not appear in the img-src directive of the Content Security Policy.
CONSOLE MESSAGE: [Report Only] Refused to load ftp://blah.test/ because it does not appear in the img-src directive of the Content Security Policy.
CONSOLE MESSAGE: [Report Only] Refused to load ftp://blah.test/ because it does not appear in the img-src directive of the Content Security Policy.


PASS 1-Violation report status OK
PASS 2-Violation report status OK
